fix: status-guard MarkRetry and split lease tests
Guard Queue.MarkRetry on status=running so concurrent CancelByLoop cannot
be resurrected to queued; treat zero-row updates as already-finalized.
Split operation_lease_test coverage by admit/bind, stop race, and finalize
contracts to stay under the AGENTS.md test-file growth threshold.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
