+++ internal/storage/repositories_test.go
@@ -1563,6 +1563,35 @@ func TestQueueRetryFailCompleteTransitions(t *testing.T) {
 		t.Fatalf("Queue.GetByID(qi_retry) after markRetry = %#v, want queued attempts=2 unclaimed", gotRetry)
 	}
 
+	// MarkRetry must not resurrect a terminal cancellation (status-guarded on running).
+	if err := repos.Queue.Upsert(ctx, QueueItemRecord{
+		ID:          "qi_retry_cancelled",
+		LoopID:      &loopID,
+		Type:        "planner",
+		TargetType:  "project",
+		TargetID:    "project_tr",
+		DedupeKey:   "d_retry_cancelled",
+		Priority:    1,
+		Status:      "cancelled",
+		AvailableAt: now,
+		Attempts:    1,
+		MaxAttempts: 3,
+		CreatedAt:   now,
+		UpdatedAt:   claimedAt,
+	}); err != nil {
+		t.Fatalf("Queue.Upsert(qi_retry_cancelled) error = %v", err)
+	}
+	if err := repos.Queue.MarkRetry(ctx, QueueMarkRetryInput{ID: "qi_retry_cancelled", AvailableAt: retryAt, Attempts: 2, ErrorMessage: &errMsg, ErrorKind: "retryable_transient", UpdatedAt: retryAt}); err != nil {
+		t.Fatalf("Queue.MarkRetry(cancelled) error = %v", err)
+	}
+	gotCancelled, err := repos.Queue.GetByID(ctx, "qi_retry_cancelled")
+	if err != nil {
+		t.Fatalf("Queue.GetByID(qi_retry_cancelled) error = %v", err)
+	}
+	if gotCancelled == nil || gotCancelled.Status != "cancelled" {
+		t.Fatalf("Queue.MarkRetry on cancelled = %#v, want status still cancelled (zero-row no-op)", gotCancelled)
+	}
+
 	finished := "2026-04-11T12:06:00.000Z"
 	if err := repos.Queue.Complete(ctx, "qi_retry", finished); err != nil {
 		t.Fatalf("Queue.Complete() error = %v", err)
